Move all tracing stuff to a separate file to make reuse easier.
